refactor: separate business logic from Button component using a custom hook
