+++ src/components/article/hooks/useHandleDeleteComment.ts
@@ -1,0 +1,20 @@
+'use client';
+
+import { deleteComment } from '@/api';
+import { useAuthStore } from '@/stores/authStore';
+
+function useHandleDeleteComment(slug: string) {
+    const { token, userInfo } = useAuthStore();
+
+    const handleDeleteComment = async (commentId: number) => {
+        try {
+            await deleteComment({ slug, commentId, token });
+        } catch (error) {
+            console.error(`Error : ${error}`);
+        }
+    };
+
+    return { userInfo, handleDeleteComment };
+}
+
+export { useHandleDeleteComment };
